FEAT

+++ src/Pages.js
@@ -25,19 +25,22 @@ function Pages() {
   const location = useLocation();
 
   useEffect(() => {
+      console.log(sessionStorage.token)
     const getTokenData = async () => {
-      try {
         const res = await axios.post(
-          "https://h8viqjk6ob.execute-api.us-east-2.amazonaws.com/default/2023-c-capstone-login",
-          {
-            type: "verify",
-            token: sessionStorage.token,
-          },
+            "https://h8viqjk6ob.execute-api.us-east-2.amazonaws.com/default/2023-c-capstone-login",
+            {
+                type: "verify",
+                token: sessionStorage.token,
+            },
         );
-      } catch (e) {
-        sessionStorage.clear();
-        window.location.reload();
-      }
+        if(res.status === 200) {
+            console.log('token verified')
+        }
+        else {
+            sessionStorage.clear();
+            window.location.reload();
+        }
     };
     if (sessionStorage.token !== undefined) {
       getTokenData();
